Extract status enum to separate concern
